requirements.txt
langchain
langchain-openai
python-dotenv==1.0.1
pydantic